code/llm_service/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pika
sqlalchemy
psycopg2-binary
pgvector
google-generativeai
openai
